omo PC don spoil

+++ src/services/deleteMovies.js
@@ -9,7 +9,7 @@ const __dirname = dirname(__filename);
 
 const deleteExpiredMovies = async () => {
     try {
-        const result = await pool.query('SELECT file_path,cover_img FROM movies WHERE expiry_date < NOW()')
+        const result = await pool.query('SELECT file_id, file_path, cover_img FROM movies WHERE expiry_date <= NOW()')
         if (result.rows.length === 0) return;
         for (let row of result.rows) {
             const files = [row.file_path, row.cover_img]
@@ -18,15 +18,20 @@ const deleteExpiredMovies = async () => {
                 if (!file) continue;
                 const filePath = path.join(process.cwd(), file);
                 try {
-                    fs.unlinkSync(filePath)
-                    console.log(`${filePath} Deleted`)
+                    if (fs.existsSync(filePath)) {
+                        fs.unlinkSync(filePath);
+                        console.log(`${filePath} Deleted`);
+                    } else {
+                        console.log(`${filePath} already gone, skipping...`);
+                    }
                 }
                 catch (err) {
                     console.error(err)
                 }
+
             }
+            await pool.query('DELETE FROM movies WHERE file_id = $1', [row.file_id])
         }
-        await pool.query('DELETE FROM movies WHERE expiry_date < NOW()')
         await redisClient.del('movies_list');
         console.log('Deleted from Database');
     }
